requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
jinja2>=3.1.0
python-multipart>=0.0.9
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.43.0
peft>=0.12.0
numpy>=1.24.0
scikit-learn>=1.4.0
huggingface_hub>=0.23.0
grad-cam>=1.5.0

# Professional additions
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=24.0.0
python-dotenv>=1.0.0
